ai-service/requirements.txt
torch
torchvision
numpy
pandas
pillow
fastapi
uvicorn
python-multipart
